refactor: move effective timeout
